CLEANUP: Enable extra warnings and fix some of them ADD: Slightly improved the ProgressWatcher interface, should be ready now for implementation in gui #29

+++ interface/gui/importers/VideoPreviewModel.cpp
@@ -58,7 +58,7 @@ int VideoPreviewModel::columnCount( const QModelIndex& ) const
 	
 QVariant VideoPreviewModel::data( const QModelIndex& index, int role ) const {
 	int pos = index.row()*columnCount(index) + index.column();
-	if( !index.isValid() || pos < 0 || pos >= previews.size() )
+	if( !index.isValid() || pos < 0 || (unsigned)pos >= previews.size() )
 		return QVariant();
 	
 	return previews.at( pos ).img;
